daemonize
snoop can now be run as a daemon, by running with --daemon (-d) option.
the process can be controlled with signals, most importantly `SIGUSR1`
(10), which forces snoop to reload its processors. If the new processor
defines a `load` method, this will be used to import the state of the
old processor instance.

also, taking a cue from apache, all processors are stored in
`snoop/processors_all`, and the active ones are symlinks in
`snoop/processors` that point to those. when deleting a processor
symlink, be sure to get the `.pyc` too!