processors are disabled on unhandled exception
If an exception other than `ProcessorAbort` is caught by the
`ProcessorBlock`, the process's `enabled` flag is set `False`.

Disabled processors are neither called per event nor sampled. By
default, processors are re-enabled upon reload, but this can be
explicitly overridden by setting `enabled = False` in the processor's
`load` method.

The `ProcessorAbort` exception should be raised to signal a non-critical
condition that requires the processor to bail on the current event, but
keep the processor enabled.